# src/infrastructure/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Product API
  description: API para gerenciamento de produtos
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Servidor local

paths:
  /products:
    get:
      summary: List all products
      operationId: getProducts
      tags:
        - Products
      responses:
        '200':
          description: Successfully retrieved products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      summary: Add new product
      operationId: addProduct
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Successfully created products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{id}:
    put:
      summary: Update an existing product
      operationId: updateProduct
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Successfully updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      summary: Delete product
      operationId: deleteProducts
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Successfully delete product
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'



components:
  schemas:
    Product:
      type: object
      properties:
        name:
          type: string
          example: "Product Name"
        price:
          type: number
          format: float
          example: 12.90